Add components compose compat module
COAND-750
