+++ components-compose/build.gradle
@@ -12,14 +12,14 @@ plugins {
 }
 
 // Maven artifact
-ext.mavenArtifactId = "drop-in-compose"
-ext.mavenArtifactName = "Adyen checkout drop-in component compose"
-ext.mavenArtifactDescription = "Compose compat Adyen checkout drop-in component."
+ext.mavenArtifactId = "components-compose"
+ext.mavenArtifactName = "Adyen checkout components compose"
+ext.mavenArtifactDescription = "Compose compat Adyen checkout components."
 
 apply from: "${rootDir}/config/gradle/sharedTasks.gradle"
 
 android {
-    namespace 'com.adyen.checkout.dropin.compose'
+    namespace 'com.adyen.checkout.components.compose'
     compileSdkVersion compile_sdk_version
 
     defaultConfig {
@@ -43,8 +43,10 @@ android {
 
 dependencies {
     // Checkout
-    api project(':drop-in')
+    api project(':components-core')
+    api project(':sessions-core')
+    api project(':ui-core')
 
     implementation platform(libraries.compose.bom)
-    implementation libraries.compose.activity
+    implementation libraries.compose.viewmodel
 }
