+++ gradle/verification-metadata.xml
@@ -49,6 +49,14 @@
             <sha256 value="b561e41c1bddb2160f79b9bc7541ad9ed110147ab2ccf63719a23d498470b043" origin="Generated by Gradle"/>
          </artifact>
       </component>
+      <component group="androidx.activity" name="activity" version="1.7.0">
+         <artifact name="activity-1.7.0.aar">
+            <sha256 value="e44b2032273387156982912c591ef7e2dd485ba0b2e689b528b5a42f271a4f27" origin="Generated by Gradle"/>
+         </artifact>
+         <artifact name="activity-1.7.0.module">
+            <sha256 value="2a746b0126a8ab2f576e7167f1a785b452ef7ee997abd979ee0c5a29c36f6ea6" origin="Generated by Gradle"/>
+         </artifact>
+      </component>
       <component group="androidx.activity" name="activity" version="1.7.1">
          <artifact name="activity-1.7.1.aar">
             <sha256 value="3b72323c5f7b6c85fa33dc541944d0905124b74675c6d0c01ea63b9029d09f1c" origin="Generated by Gradle"/>
@@ -105,6 +113,14 @@
          </artifact>
          <artifact name="activity-ktx-1.6.0.module">
             <sha256 value="574b89ee589c530ec4d3fc412aac890066842db45a6a9a1bbe89542e0b47e81b" origin="Generated by Gradle"/>
+         </artifact>
+      </component>
+      <component group="androidx.activity" name="activity-ktx" version="1.7.0">
+         <artifact name="activity-ktx-1.7.0.aar">
+            <sha256 value="fce317d61a22f12967b475bfcb80c89dda66e418975e890ea703cb74e12b5b11" origin="Generated by Gradle"/>
+         </artifact>
+         <artifact name="activity-ktx-1.7.0.module">
+            <sha256 value="f4001a709b6f7132b22a5e17c24f8b8a9b3c462a598de3b0f16cc57a4344bab6" origin="Generated by Gradle"/>
          </artifact>
       </component>
       <component group="androidx.activity" name="activity-ktx" version="1.7.1">
