+++ settings.gradle
@@ -25,6 +25,7 @@ include ':3ds2',
         ':boleto',
         ':card',
         ':checkout-core',
+        ':components-compose',
         ':components-core',
         ':convenience-stores-jp',
         ':cse',
